Refactor: extract shared modules to eliminate code duplication
- Create constants.py for magic strings (AE_ENTRY_TEMPLATE, ENTRY_LOGID_RE, HEADER_MARK, FOOTER_MARK)
- Create topics.py with TopicDetector class and load_topics() function
- Create entries.py with build_action_string() and format_entry() functions
- Refactor backfill_pending_changes.py to use shared modules (~47% reduction)
- Remove unused mediawiki_param_nowiki() function from bot.py
- Update test imports for new module structure

This eliminates ~150 lines of duplicated code between bot.py and backfill_pending_changes.py,
improves maintainability by centralizing magic strings, and follows single responsibility principle.
